requirements.txt
streamlit
python-dotenv
PyMuPDF
pdf2image
pytesseract
google-generativeai